New themes! (LiuChan and Pleco)
Pleco imitates the Pleco Dictionary for Android's theme.
LiuChan is the new main theme.

Changed the dictionary parser's serializer code to improve
the HTML markup.

All older themes have been updated and purged from defunct code.

Added test.html page for easy changing of theme css
(note: this is not meant to be packed with the extension on deployment)