Reworked dictionary handling code. Dictionary file has been updated to version 2017-09-02 with 115480 entries.
Extension size has been reduced by about 3mb.

The extension now directly loads CEDict as an object (as opposed to
storing it in memory as a mere string.

Furthermore the code first iterates over the dictionary and gets an
index of entries that start with the first character in the word.

CPU usage has been reduced by about 10-7%, the biggest factor in this
was onmousemove firing way too often (now on 100ms delay).