v1.0.3 Release
Added Theme customization
- You can now alter the tone colors of any theme, to ensure you don't
have to re-associate the colors you're used to!

The options page has been revamped somewhat to accommodate the theme
customization - allowing on the fly editing of themes, plus the addition
of color pickers (courtesy of color-picker.js by Taufik Nurrohman).

LiuChan will no longer highlight text when an input is active,
regardless of settings! This is to prevent inputs from losing focus if
you move your move while typing. Furthermore, hotkeys are disabled when
an input is active to prevent annoyances.

Added various optimizations.
-Content scripts now only hold the options they need access to instead
of all of them

-Dictionary is now rebuilt if the user selects a different pinyin
type, so that the extension does not have to disabled and enabled again
for effects to take hold.

Popup Delay has been fixed