Massively improved performance by switching to Map. As an added side-effect variant type character matches are now also included in results and memory usage is down by 50%+. Testing showed that processing 100000 entries has been reduced to 6 sec as opposed to 250 seconds!
Liuchan now also only does its character lookup using simplified
characters, essentially cutting the cycles necessary in half.

Altered fuzzysort to make use of the map, but it's still unoptimized
and messy.

When trying to enable Liuchan on a tab that doesn't have a content
script running, it will now notify the user to reload the tab.