v1.1.4 Release
Fixed a fairly serious bug that affected only new users of the app.
Didn't catch this before because I didn't clear my synchronized settings
for a long time...

Fixed a bug where if a page did some funky dynamic loading and unloaded
the popup element, it would not create a new one.